+++ CHANGELOG.md
@@ -13,8 +13,13 @@
 
 ## [1.0.0] - 2022-06-08
 
+#### New Features:
+* Subject attributes: an optional `subject_attributes` param is added to the `get_assignment` function. The subject attributes may contains custom metadata about the subject. These attributes are used for evaluating any targeting rules defined on the experiment.
+```
+client.get_assignment("<SUBJECT_KEY">, "<EXPERIMENT_KEY>", { "email": "user@example.com" });
+```
+
 #### Breaking Changes:
-* Subject attributes: the `subject` parameter of the assignment function was changed from a string to an object. The new `subject` object contains a `key` field for the subject ID as well as an optional `custom_attributes` property for any related metadata like name or email.
 * The EppoClient `assign()` function is renamed to `get_assignment()`
 
 ## [0.0.3] - 2022-05-11